Port to Windows (MinGW GCC 13.1 + Qt 6.9.1)
Add full Windows build support using platform-conditional compilation.
All changes are guarded by PLATFORM_WINDOWS / WIN32 so the Linux build
is unaffected.

New platform files: EmCommonWin.h, Platform_Win.cpp, EmDirRefWin.cpp,
EmFileRefWin.cpp, EmPixMapWin.cpp, EmDocumentWin.cpp,
EmTransportSerialWin.cpp (stubbed).

Fix LLP64 pointer-to-integer casts (long → intptr_t) in EmMemory.h,
Platform_NetLib_Sck.cpp, Logging.cpp, and EmPatchModuleNetLib.cpp.

Add Winsock2 includes for socket code, Windows endian detection in
Switches.h, and NSIS installer script.

Includes Windows icon (pose64.ico) and resource file (pose64.rc) for
exe branding and version info, plus mkico.c build tool.

+++ src/core/Hardware/EmMemory.h
@@ -220,7 +220,7 @@ STATIC_INLINE uint16 EmMemDoGet16 (void* a)
 STATIC_INLINE uint8 EmMemDoGet8 (void* a)
 {
 #if WORDSWAP_MEMORY
-	return *(uint8*) ((long) a ^ 1);
+	return *(uint8*) ((intptr_t) a ^ 1);
 #else
 	return *(uint8*) a;
 #endif
@@ -256,7 +256,7 @@ STATIC_INLINE void EmMemDoPut16 (void* a, uint16 v)
 STATIC_INLINE void EmMemDoPut8 (void* a, uint8 v)
 {
 #if WORDSWAP_MEMORY
-	*(uint8*) ((long) a ^ 1) = v;
+	*(uint8*) ((intptr_t) a ^ 1) = v;
 #else
 	*(uint8*) a = v;
 #endif
